requirements.txt
asgiref==3.4.1
confusable-homoglyphs==3.2.0
Django==3.2.5
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
